+++ src/Exception/BadParameterException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Obullo\Router\Exception;
-
-use LengthException;
-
-class BadParameterException extends LengthException
-{
-}