+++ src/Exception/ParseException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Obullo\Router\Exception;
-
-use RuntimeException;
-
-class ParseException extends RuntimeException
-{
-}